requirements.txt
python==3.9
mysql-connector-python
ckanapi